Move report helper utils to common_report
