+++ quipucords/api/deployments_report/util.py
@@ -15,7 +15,7 @@ import csv
 import logging
 from io import StringIO
 
-from api.common.util import CSVHelper
+from api.common.common_report import CSVHelper, sanitize_row
 from api.models import (DeploymentsReport,
                         Source,
                         SystemFingerprint)
@@ -49,12 +49,6 @@ def compute_source_info(sources):
             result[SATELLITE_DETECTION_KEY] = True
         result[SOURCES_KEY].append(source.get('source_name'))
     return result
-
-
-def sanitize_row(row):
-    """Replace commas in fact values to prevent false csv parsing."""
-    return [fact.replace(',', ';')
-            if isinstance(fact, str) else fact for fact in row]
 
 
 def create_deployments_csv(deployments_report_dict):
