+++ quipucords/api/reports/reports_gzip_renderer.py
@@ -13,7 +13,7 @@
 import logging
 
 import api.messages as messages
-from api.common.util import (create_filename, create_tar_buffer)
+from api.common.common_report import (create_filename, create_tar_buffer)
 from api.deployments_report.util import create_deployments_csv
 from api.details_report.util import create_details_csv
